requirements.txt
python-dotenv
pandas
numpy
pymongo[srv]>=4.5
pymongo
certifi
black
isort
pre-commit

-e .